Feature/task 9

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.LendingService;
 
 /**
  * 削除コントローラー
@@ -24,6 +25,9 @@ public class DeleteBookController {
     @Autowired
     private BooksService booksService;
 
+    @Autowired
+    private LendingService lendingService;
+
     /**
      * 対象書籍を削除する
      *
@@ -34,24 +38,28 @@ public class DeleteBookController {
      */
     @Transactional
     @RequestMapping(value = "/deleteBook", method = RequestMethod.POST) //RequestMappng とはSpring MVC のコントローラに付与して、
-                                                                        //リクエスト URL に対して、どのメソッドが処理を実行するか定義するアノテーション。
+                                                                                       //リクエスト URL に対して、どのメソッドが処理を実行するか定義するアノテーション。
     public String deleteBook(
             Locale locale,
             @RequestParam("bookId") Integer bookId,
             Model model) {
         logger.info("Welcome delete! The client locale is {}.", locale);
-
-
-        //削除メソッドを使用する
-        booksService.deleteBook(bookId);
-
-        //新しい書籍リストを取得する
-        model.addAttribute("bookList", booksService.getBookList());
-
-        //ホーム画面に遷移する
-        return "home";
-        
-
+        //貸出状況の確認
+        int count = lendingService.LendingConfirmation(bookId);
+        if (count == 0) {
+            //削除メソッドを使用する
+            booksService.deleteBook(bookId);
+            //新しい書籍リストを取得する
+            model.addAttribute("bookList", booksService.getBookList());
+            //ホーム画面に遷移する
+            return "home";
+        } else {
+            model.addAttribute("errorDelete", "貸出中のため削除できません");
+            //書籍情報を再取得する
+            model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+            //貸出ステータス
+            model.addAttribute("lendingStatus", "貸出中");
+            return "details";
+        }
     }
-
 }
